requirements.txt
geopy==2.2.0
pandas==1.4.1
psycopg2==2.9.3
psycopg2_binary==2.9.3
pytest==6.0.1
PyYAML==6.0
requests==2.24.0
